Split out tracing tests
