+++ smallrye-reactive-messaging-amqp/src/test/java/io/smallrye/reactive/messaging/amqp/TracingAmqpToAppNoParentTest.java
@@ -1,0 +1,151 @@
+package io.smallrye.reactive.messaging.amqp;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.awaitility.Awaitility.await;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.CompletionStage;
+import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.concurrent.atomic.AtomicInteger;
+
+import javax.enterprise.context.ApplicationScoped;
+
+import org.eclipse.microprofile.config.ConfigProvider;
+import org.eclipse.microprofile.reactive.messaging.Incoming;
+import org.eclipse.microprofile.reactive.messaging.Message;
+import org.jboss.weld.environment.se.Weld;
+import org.jboss.weld.environment.se.WeldContainer;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import io.opentelemetry.api.GlobalOpenTelemetry;
+import io.opentelemetry.api.trace.Span;
+import io.opentelemetry.api.trace.SpanKind;
+import io.opentelemetry.api.trace.propagation.W3CTraceContextPropagator;
+import io.opentelemetry.context.propagation.ContextPropagators;
+import io.opentelemetry.sdk.OpenTelemetrySdk;
+import io.opentelemetry.sdk.testing.exporter.InMemorySpanExporter;
+import io.opentelemetry.sdk.trace.SdkTracerProvider;
+import io.opentelemetry.sdk.trace.SpanProcessor;
+import io.opentelemetry.sdk.trace.data.SpanData;
+import io.opentelemetry.sdk.trace.export.SimpleSpanProcessor;
+import io.opentelemetry.sdk.trace.samplers.Sampler;
+import io.smallrye.config.SmallRyeConfigProviderResolver;
+import io.smallrye.reactive.messaging.TracingMetadata;
+import io.smallrye.reactive.messaging.test.common.config.MapBasedConfig;
+
+public class TracingAmqpToAppNoParentTest extends AmqpBrokerTestBase {
+
+    private InMemorySpanExporter testExporter;
+    private SpanProcessor spanProcessor;
+
+    private WeldContainer container;
+    private final Weld weld = new Weld();
+
+    @BeforeEach
+    public void setup() {
+        super.setup();
+        GlobalOpenTelemetry.resetForTest();
+
+        testExporter = InMemorySpanExporter.create();
+        spanProcessor = SimpleSpanProcessor.create(testExporter);
+
+        SdkTracerProvider tracerProvider = SdkTracerProvider.builder()
+                .addSpanProcessor(spanProcessor)
+                .setSampler(Sampler.alwaysOn())
+                .build();
+
+        OpenTelemetrySdk.builder()
+                .setPropagators(ContextPropagators.create(W3CTraceContextPropagator.getInstance()))
+                .setTracerProvider(tracerProvider)
+                .buildAndRegisterGlobal();
+    }
+
+    @AfterEach
+    public void cleanup() {
+        if (container != null) {
+            container.close();
+        }
+        // Release the config objects
+        SmallRyeConfigProviderResolver.instance().releaseConfig(ConfigProvider.getConfig());
+
+        if (testExporter != null) {
+            testExporter.shutdown();
+        }
+        if (spanProcessor != null) {
+            spanProcessor.shutdown();
+        }
+    }
+
+    @AfterAll
+    static void shutdown() {
+        GlobalOpenTelemetry.resetForTest();
+    }
+
+    @Test
+    public void testFromAmqpToAppWithNoParent() {
+        weld.addBeanClass(MyAppReceivingData.class);
+
+        new MapBasedConfig()
+                .put("mp.messaging.incoming.stuff.connector", AmqpConnector.CONNECTOR_NAME)
+                .put("mp.messaging.incoming.stuff.host", host)
+                .put("mp.messaging.incoming.stuff.port", port)
+                .put("mp.messaging.incoming.stuff.address", "no-parent-stuff")
+
+                .put("amqp-username", username)
+                .put("amqp-password", password)
+                .write();
+
+        container = weld.initialize();
+        MyAppReceivingData bean = container.getBeanManager().createInstance().select(MyAppReceivingData.class).get();
+
+        AtomicInteger count = new AtomicInteger();
+
+        usage.produce("no-parent-stuff", 10, count::getAndIncrement);
+
+        await().until(() -> bean.list().size() >= 10);
+        assertThat(bean.list()).containsExactly(0, 1, 2, 3, 4, 5, 6, 7, 8, 9);
+
+        assertThat(bean.tracing()).hasSizeGreaterThanOrEqualTo(10);
+        assertThat(bean.tracing()).doesNotContainNull().doesNotHaveDuplicates();
+        List<String> spanIds = new ArrayList<>();
+
+        for (TracingMetadata tracing : bean.tracing()) {
+            spanIds.add(tracing.getCurrentSpanContext().getSpanId());
+            assertThat(Span.fromContextOrNull(tracing.getPreviousContext())).isNull();
+        }
+
+        assertThat(spanIds).doesNotContainNull().doesNotHaveDuplicates().hasSizeGreaterThanOrEqualTo(10);
+
+        for (SpanData data : testExporter.getFinishedSpanItems()) {
+            assertThat(data.getSpanId()).isIn(spanIds);
+            assertThat(data.getSpanId()).isNotEqualTo(data.getParentSpanId());
+            assertThat(data.getKind()).isEqualByComparingTo(SpanKind.CONSUMER);
+        }
+    }
+
+    @ApplicationScoped
+    public static class MyAppReceivingData {
+        private final List<TracingMetadata> tracingMetadata = new ArrayList<>();
+        private final List<Integer> results = new CopyOnWriteArrayList<>();
+
+        @Incoming("stuff")
+        public CompletionStage<Void> consume(Message<Integer> input) {
+            results.add(input.getPayload());
+            tracingMetadata.add(input.getMetadata(TracingMetadata.class).orElse(TracingMetadata.empty()));
+            return input.ack();
+        }
+
+        public List<Integer> list() {
+            return results;
+        }
+
+        public List<TracingMetadata> tracing() {
+            return tracingMetadata;
+        }
+    }
+
+}
